refactor: add commands integration test file
